use bfloat16

+++ skyrl-tx/tests/models/test_qwen3_generate.py
@@ -87,7 +87,7 @@ def test_qwen3_generate_speed():
         hf_model.save_pretrained(tmp, safe_serialization=True)
         mesh = jax.make_mesh((1, 1), ("dp", "tp"))
         with jax.set_mesh(mesh):
-            model = Qwen3ForCausalLM(config, dtype=jnp.float32, rngs=nnx.Rngs(0))
+            model = Qwen3ForCausalLM(config, dtype=jnp.bfloat16, rngs=nnx.Rngs(0))
         load_safetensors(tmp, config, model)
 
         # Warmup
